Release v2.0.0 (#12)
* Release v2.0.0

Version 1.0.0 had largely the same behavior that development versions did.  In particular,
`ilog2(x)` threw a `DomainError` for most integer types with `x` less than or equal to
zero. But other types, such as `Float64` did not cause a `DomainError` to be thrown, and instead
returned `ilog2(abs(x))`.

Version `2.0.0` extends the behavior for integers to all types. Now there is a single entry point
to `ilog2` that makes this check.

* Fix test for Julia 1.5 and earlier

One test is slightly different for Julia 1.5 and earlier